added TargetAudience context methods

+++ IntelligentSampleEnginePOC.API/IntelligentSampleEnginePOC.API.Core/Interfaces/IProjectContext.cs
@@ -15,6 +15,6 @@ namespace IntelligentSampleEnginePOC.API.Core.Interfaces
 
         ProjectList GetProjects(int? status,int pageNumber, string? searchString, int recentCount);
 
-        Project? Get(int id);
+        Project? Get(long id);
     }
 }
